flask/requirements.txt
cfgv==3.4.0
distlib==0.3.8
filelock==3.15.4
identify==2.6.0
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.8.0
PyYAML==6.0.1
virtualenv==20.26.3
Flask==2.2.5
Werkzeug==2.2.2
MarkupSafe==2.1.1
opencv-python==4.10.0.84
pillow==10.4.0
aniso8601==9.0.1
apispec==6.4.0
apispec-webframeworks==1.0.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
bidict==0.23.1
binaryornot==0.4.4
blinker==1.8.2
boto3==1.34.34
botocore==1.34.162
certifi==2024.8.30
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cookiecutter==2.5.0
deepdiff==6.7.1
distlib==0.3.8
Django==5.1.1
dnspython==2.7.0
et-xmlfile==1.1.0
filelock==3.15.4
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
gevent==23.9.1
gevent-websocket==0.10.1
gitignore_parser==0.1.11
greenlet==3.1.1
h11==0.14.0
hyperlink==21.0.0
identify==2.6.0
idna==3.10
image==1.5.33
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kthread==0.2.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.20.2
mdurl==0.1.2
networkx==3.2.1
nodeenv==1.9.1
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.2
ordered-set==4.1.0
packaging==24.1
pandas==2.2.0
passlib==1.7.4
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.8.0
pyarrow==15.0.0
Pygments==2.18.0
pymongo==4.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-slugify==8.0.4
python-socketio==5.11.4
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.32.3
rich==13.9.2
s3transfer==0.10.2
setuptools==75.1.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.5.1
taipy==3.1.1
taipy-config==3.1.1
taipy-core==3.1.1
taipy-gui==3.1.4
taipy-rest==3.1.1
taipy-templates==3.1.1
text-unidecode==1.3
toml==0.10.2
Twisted==23.10.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
virtualenv==20.26.3
Werkzeug==3.0.4
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.3
